pyproject.toml
[tool.poetry]
name = "py-ai-certificate"
version = "0.1.0"
description = "Selection process for The Cliff"
authors = ["Ivan Hernandez <ivanhernandezroldan1@gmail.com"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
streamlit = "^1.30.0"
langchain = "^0.1.2"
openai = "^1.9.0"
chromadb = "^0.4.22"
st-pages = "^0.4.5"
unstructured = "^0.12.2"
markdown = "^3.5.2"
pypdf = "^4.0.0"
python-dotenv = "^1.0.1"
langchain-openai = "^0.0.3"
pypdf2 = "^3.0.1"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
